+++ src/main/java/api/box/Application.java
@@ -1,4 +1,4 @@
-package api.hello;
+package api.box;
 
 import org.springframework.boot.SpringApplication;
 import org.springframework.boot.autoconfigure.SpringBootApplication;
