--- src/main/java/api/hello/GreetingController.java
@@ -1,24 +1,0 @@
-package api.hello;
-
-import java.util.concurrent.atomic.AtomicLong;
-import org.springframework.web.bind.annotation.RequestMapping;
-import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.RestController;
-import org.w3c.dom.css.Counter;
-
-/*
- * @RestControllerはドメインオブジェクトを返すコントローラーとしてのクラスを表す
- * @Controllerと@ResponseBodyをまとめてロールしたもの
- */
-@RestController
-public class GreetingController {
-	private static final String template = "Hello, %s";
-	private final AtomicLong conter = new AtomicLong();
-	
-	@RequestMapping("/greeting")
-	public Greeting greeting(@RequestParam(value="name", defaultValue="World")String name) {
-		return new Greeting(conter.incrementAndGet(), String.format(template, name));
-	}
-	
-
-}
